requirements.txt

opencv-python==4.6.0.66
pytesseract==0.3.10
Pillow==9.0.1
pandas==1.4.2
numpy==1.21.5
scipy==1.7.3